+++ crates/board/src/crypto/aead.rs
@@ -18,7 +18,7 @@ use generic_array::{ArrayLength, GenericArray};
 #[cfg(feature = "internal-aead")]
 pub use software::*;
 
-use crate::{Error, Support, Unsupported};
+use crate::{Error, Support};
 
 #[derive(Copy, Clone)]
 pub struct AeadSupport {
@@ -33,19 +33,21 @@ impl From<AeadSupport> for bool {
 }
 
 /// Elliptic-curve cryptography interface.
-pub trait Api<Key, Iv, Tag>: Support<AeadSupport>
+pub trait Api<Key, Iv>: Support<AeadSupport>
 where
     Key: ArrayLength<u8>,
     Iv: ArrayLength<u8>,
-    Tag: ArrayLength<u8>,
 {
+    /// The tag length.
+    type Tag: ArrayLength<u8>;
+
     /// Encrypts and authenticates a clear text with associated data given a key and IV.
     ///
     /// The clear- and cipher-texts must have the same length. If the clear text is omitted, then
-    /// the cipher text is encrypted in place.
+    /// the cipher text is encrypted in place. The tag must be `TAG_LEN` bytes long.
     fn encrypt(
         key: &Array<Key>, iv: &Array<Iv>, aad: &[u8], clear: Option<&[u8]>, cipher: &mut [u8],
-        tag: &mut Array<Tag>,
+        tag: &mut Array<Self::Tag>,
     ) -> Result<(), Error>;
 
     /// Decrypts and authenticates a cipher text with associated data given a key and IV.
@@ -53,31 +55,39 @@ where
     /// The cipher- and clear-texts must have the same length. If the cipher text is omitted, then
     /// the clear text is decrypted in place.
     fn decrypt(
-        key: &Array<Key>, iv: &Array<Iv>, aad: &[u8], cipher: Option<&[u8]>, tag: &Array<Tag>,
-        clear: &mut [u8],
+        key: &Array<Key>, iv: &Array<Iv>, aad: &[u8], cipher: Option<&[u8]>,
+        tag: &Array<Self::Tag>, clear: &mut [u8],
     ) -> Result<(), Error>;
 }
 
 pub type Array<N> = GenericArray<u8, N>;
 
-impl Support<AeadSupport> for Unsupported {
+pub struct Unsupported<Tag: ArrayLength<u8>> {
+    _never: !,
+    _tag: Tag,
+}
+
+impl<Tag: ArrayLength<u8>> Support<AeadSupport> for Unsupported<Tag> {
     const SUPPORT: AeadSupport = AeadSupport { no_copy: false, in_place_no_copy: false };
 }
 
-impl<Key, Iv, Tag> Api<Key, Iv, Tag> for Unsupported
+impl<Key, Iv, Tag: ArrayLength<u8>> Api<Key, Iv> for Unsupported<Tag>
 where
     Key: ArrayLength<u8>,
     Iv: ArrayLength<u8>,
-    Tag: ArrayLength<u8>,
 {
+    type Tag = Tag;
+
     fn encrypt(
-        _: &Array<Key>, _: &Array<Iv>, _: &[u8], _: Option<&[u8]>, _: &mut [u8], _: &mut Array<Tag>,
+        _: &Array<Key>, _: &Array<Iv>, _: &[u8], _: Option<&[u8]>, _: &mut [u8],
+        _: &mut Array<Self::Tag>,
     ) -> Result<(), Error> {
         unreachable!()
     }
 
     fn decrypt(
-        _: &Array<Key>, _: &Array<Iv>, _: &[u8], _: Option<&[u8]>, _: &Array<Tag>, _: &mut [u8],
+        _: &Array<Key>, _: &Array<Iv>, _: &[u8], _: Option<&[u8]>, _: &Array<Self::Tag>,
+        _: &mut [u8],
     ) -> Result<(), Error> {
         unreachable!()
     }
@@ -94,18 +104,19 @@ mod software {
         const SUPPORT: AeadSupport = AeadSupport { no_copy: false, in_place_no_copy: true };
     }
 
-    impl<Key, Iv, Tag, T> Api<Key, Iv, Tag> for T
+    impl<Key, Iv, T> Api<Key, Iv> for T
     where
         T: KeyInit + AeadInPlace,
         T: KeySizeUser<KeySize = Key>,
-        T: AeadCore<NonceSize = Iv, TagSize = Tag>,
+        T: AeadCore<NonceSize = Iv>,
         Key: ArrayLength<u8>,
         Iv: ArrayLength<u8>,
-        Tag: ArrayLength<u8>,
     {
+        type Tag = T::TagSize;
+
         fn encrypt(
             key: &Array<Key>, iv: &Array<Iv>, aad: &[u8], clear: Option<&[u8]>, cipher: &mut [u8],
-            tag: &mut Array<Tag>,
+            tag: &mut Array<Self::Tag>,
         ) -> Result<(), Error> {
             let aead = T::new(key);
             if let Some(clear) = clear {
@@ -118,8 +129,8 @@ mod software {
         }
 
         fn decrypt(
-            key: &Array<Key>, iv: &Array<Iv>, aad: &[u8], cipher: Option<&[u8]>, tag: &Array<Tag>,
-            clear: &mut [u8],
+            key: &Array<Key>, iv: &Array<Iv>, aad: &[u8], cipher: Option<&[u8]>,
+            tag: &Array<Self::Tag>, clear: &mut [u8],
         ) -> Result<(), Error> {
             let aead = T::new(key);
             if let Some(cipher) = cipher {